get_pipeline_report

Return your pipeline report information from Less annoying CRM.

get_pipeline_report(user_code, api_token, pipelineid)

Arguments

user_code
The user code to identify your account

api_token
The api token to connect to your account

pipelineid
The id of the pipeline you want to get the report from

Details

For this you will need to know the pipelineId, StatusId, and CustomFieldId You can get this PipelineId's at https://www.lessannoyingcrm.com/app/Settings/Api'

Examples

if (FALSE) {
get_pipeline_report(user_code = "6A6E88",
                    api_token = "96066",
                    pipelineid = "57102821")
}